Creator method works.
Creator method fully works. Still need to add support for delay and
other options. Need to fix Start/Stop button bug. Need to add
username:proxy support on account creator page. Made a tab for Username
Generator which will be used to generate usernames and tie them to
proxies that we have in the system.

Still need to add:
-Account verification.
-Gig poster
-Account checker
-Sales monitor